04A_gRPC/server/requirements.txt
grpcio-tools==1.54.0